refactor: reorganize frontend utilities
