+++ src/extension.ts
@@ -6,7 +6,7 @@ import * as logger from './logger/logUtils';
 import { initializeSecrets } from './utils/secretUtils';
 import { copyDefaultAgents, configureLatexSettings } from './utils/setupUtils';
 import { watchConfig } from './utils/config';
-import { disposeDiffRefresh } from './utils/diffViewUtils';
+import { disposeDiffRefresh } from './frontend/ui/diffView';
 
 // Local imports - components
 import { ProgressViewProvider } from './progressView/ProgressViewProvider';
