+++ src/frontend/latex/openBuild.ts
@@ -2,8 +2,8 @@
 import * as vscode from 'vscode';
 
 // Local imports - log
-import * as logger from '../logger/logUtils';
-import { fileExists, getFullPathFromWorkspace } from './files';
+import * as logger from '../../logger/logUtils';
+import { fileExists, getFullPathFromWorkspace } from '../../utils/files';
 
 const CHANNEL = 'OpenBuildUtils';
 logger.initialize(CHANNEL);
